backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
attr
python-jose[cryptography]
jwt
python-dotenv==1.0.1
pydantic==1.10.18
python-multipart
alembic==1.13.2
bcrypt==4.0.0
openpyxl==3.1.5
pydantic[email]
Pillow